requirements.txt
streamlit==1.25.0
Pillow==9.4.0
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
xgboost==1.7.6
statsmodels==0.14.0
streamlit_option_menu==0.3.12

            